fix(lint): clear the clippy backlog the revived gate exposed
Second wave of fallout from reviving Rust CI. With `clippy --all-targets -- -D warnings`
running for the first time, the workspace failed crate after crate. All fixed here;
`cargo clippy --all-targets -- -D warnings` now exits 0.

Mechanical, behaviour-preserving:
- `derivable_impls`      — `impl Default for Visibility` -> `#[derive(Default)]` + `#[default]`
- `manual_try_fold` (x2) — `.fold(Ok(acc), ..)` -> `.try_fold(acc, ..)` in both the surface
                           and core parsers' statement-sequence desugaring
- `collapsible_if` (x2), `match_like_matches_macro` — via `cargo clippy --fix`
- `ptr_arg`              — `&PathBuf` -> `&Path` in two ephapax-cli signatures
- `unnecessary_unsafe`   — removed an EMPTY `unsafe {}` block that wrapped only comments
                           in a placeholder test (`ephapax-runtime::list`). Worth noting in
                           a language whose thesis is memory safety.

Deliberately allowed, each with an inline reason rather than a blanket crate-level allow:
- `approx_constant`      — `lex("3.14")` is a float literal UNDER TEST, not an approximation
                           of PI
- `type_complexity`      — introduced a `TokenPredicate` type alias instead
- `too_many_arguments` (x3) — 8 params against clippy's default threshold of 7, on
                           option-heavy CLI entry points and the recursive import resolver
- `large_enum_variant`   — `Value::Closure` is ~264 bytes vs ~48 for the next largest, so
                           every `Value` pays for it. The fix is to box the payload, which
                           touches every construction and match site in the INTERPRETER.
                           That is a behavioural-risk refactor and does not belong inside a
                           CI-infrastructure change, so it is allowed here with a comment
                           and tracked in #360.

Verification, all local and all green:
  cargo check --locked --all-targets   ok
  cargo clippy --all-targets -- -D warnings   exit 0
  cargo fmt --all -- --check           clean
  cargo test --workspace               285 passed, 0 failed

+++ ephapax-linear/src/tests.rs
@@ -383,7 +383,7 @@ fn p2p_single_consume_always_accepted_both_disciplines() {
         let expr = e(ExprKind::LetLin {
             name: name.clone().into(),
             ty: Some(Ty::Base(BaseTy::I32)),
-            value: Box::new(i32_lit(i as i32)),
+            value: Box::new(i32_lit(i)),
             body: Box::new(var(&name)),
         });
 
@@ -413,7 +413,7 @@ fn p2p_contraction_always_rejected_both_disciplines() {
         let expr = e(ExprKind::LetLin {
             name: name.clone().into(),
             ty: Some(Ty::Base(BaseTy::I32)),
-            value: Box::new(i32_lit(i as i32)),
+            value: Box::new(i32_lit(i)),
             body: Box::new(e(ExprKind::Pair {
                 left: Box::new(var(&name)),
                 right: Box::new(var(&name)),
@@ -462,7 +462,7 @@ fn p2p_weakening_linear_forbidden_affine_allowed() {
         let expr = e(ExprKind::LetLin {
             name: name.clone().into(),
             ty: Some(Ty::Base(BaseTy::I32)),
-            value: Box::new(i32_lit(i as i32)),
+            value: Box::new(i32_lit(i)),
             body: Box::new(unit()),
         });
 
@@ -485,7 +485,7 @@ fn p2p_weakening_linear_forbidden_affine_allowed() {
         let affine_expr = e(ExprKind::Let {
             name: name.clone().into(),
             ty: Some(Ty::Base(BaseTy::I32)),
-            value: Box::new(i32_lit(i as i32)),
+            value: Box::new(i32_lit(i)),
             body: Box::new(unit()),
         });
         let mut affine = AffineChecker::new();
@@ -510,7 +510,7 @@ fn p2p_unified_api_consistent_with_direct_checkers() {
         let good_expr = e(ExprKind::LetLin {
             name: name.clone().into(),
             ty: Some(Ty::Base(BaseTy::I32)),
-            value: Box::new(i32_lit(i as i32)),
+            value: Box::new(i32_lit(i)),
             body: Box::new(var(&name)),
         });
